requirements.txt
python-binance
websocket-client
pandas
awaitable